perf(kvbit): drop 3 redundant fp32 back-casts (graph-size lever)
Graph-off trace A/B (REPORT §10) shows kvbit issues 4685 kernels/forward vs
native 3047 (1.54x); +1326 elementwise/copy (5.1x, ~17/layer) is 80% of the
gap. The bigger CUDA graph drives slower cuGraphLaunch replay (graph-on
profile: +9ms ITL, only +0.76ms GPU). Three lossless fp32 back-cast cuts:

- set_mla_kv_buffer: pass bf16 rotated to the fused kernel (it upcasts to
  fp32 internally), drop the outer .to(torch.float32). Bit-identical quant.
- Q-FHT fold (forward_mla.py): drop .to(q_nope_out.dtype) back-cast. The
  bf16@bf16 GMMA result widened to fp32 carries no extra info (bf16->fp32 is
  lossless); downstream (kvbit decode kernel loads bf16, fa3 target_verify,
  W_vc) all consume bf16. The packgqa tl.dot wants bf16 anyway — cleaner.
- out@R inverse (dsa_backend._forward_kvbit): drop .to(out.dtype) back-cast,
  same reasoning; out flows to W_vc (bf16 linear).

Cuts 3 elementwise kernels/layer (~234/forward, ~5% of the graph). Lossless:
each back-cast only widened a bf16 matmul result to fp32. e2e A/B next to
test whether shrinking the graph closes the 30% throughput gap.

+++ python/sglang/srt/layers/attention/dsa_backend.py
@@ -1946,7 +1946,12 @@ class DeepseekSparseAttnBackend(
                 pool.kv_lora_rank, dtype=torch.bfloat16, device=out.device
             )
             self._kvbit_decode_R = R
-        out = (out.to(torch.bfloat16) @ R).to(out.dtype)
+        # Keep bf16 out of the matmul (drop the .to(out.dtype) back-cast): the
+        # bf16 @ bf16 GMMA result widened to fp32 carries no extra info, and
+        # downstream (W_vc up-projection) consumes bf16 — the back-cast was a
+        # free-floating elementwise launch inflating the CUDA graph (graph-size
+        # lever per REPORT §10). Lossless.
+        out = out.to(torch.bfloat16) @ R
         return out
 
     def _forward_flashmla_sparse(
